+++ src/assets/app.scss
@@ -368,3 +368,9 @@ h2 {
     background-color: #f8d7da;
     border-color: #f8d7da;
 }
+
+.btn-success {
+    color: #fff;
+    background-color: #4caf50;
+    border-color: #4caf50;
+}
